Initial drop of XLIFF 2.0 Object Model
This drop contains the initial implementation of the Xliff 2.0 Object
Model.  This drop includes projects for the Object Model itself
(Xliff.OM), tests for the Object Model (Xliff.OM.Tests), documentation
for the Object Model (Xliff.OM.Documentation) using Sandcastle on the C#
/// style method commenting, and samples (Xliff.Samples) that
demonstrate how to use various parts of the Object Model.

Included with this drop is a build script (build.cmd) that will build
all the projects from the command line using msbuild.exe.  Refer to
build.cmd for system requirements.